+++ modules/Grades/includes/Dashboard.inc.php
@@ -61,33 +61,10 @@ if ( ! function_exists( 'DashboardGradesAdmin' ) )
 
 		if ( ! isset( $gpa_RET[1]['CUM_WEIGHTED_GPA'] ) )
 		{
-			/**
-			 * SQL result as comma separated list
-			 *
-			 * @since 9.3 Add MySQL support
-			 * @link https://dev.mysql.com/doc/refman/5.7/en/aggregate-functions.html#function_group-concat
-			 *
-			 * @param string $column    SQL column.
-			 * @param string $separator List separator, default to comma.
-			 *
-			 * @return string MySQL or PostgreSQL function
-			 */
-			$sql_comma_separated_result = function( $column, $separator = ',' )
-			{
-				global $DatabaseType;
-
-				if ( $DatabaseType === 'mysql' )
-				{
-					return "GROUP_CONCAT(" . $column . " SEPARATOR '" . DBEscapeString( $separator ) . "')";
-				}
-
-				return "ARRAY_TO_STRING(ARRAY_AGG(" . $column . "), '" . DBEscapeString( $separator ) . "')";
-			};
-
 			// PostgreSQL version >= 8.4 required for ARRAY_TO_STRING() function.
 			// Assignments.
 			$assignments_RET = DBGet( "SELECT COUNT(ASSIGNMENT_ID) AS ASSIGNMENTS_NB,
-			" . $sql_comma_separated_result( 'ASSIGNMENT_ID' ) . " AS ASSIGNMENTS_LIST,
+			" . DBSQLCommaSeparatedResult( 'ASSIGNMENT_ID' ) . " AS ASSIGNMENTS_LIST,
 			DUE_DATE
 			FROM gradebook_assignments
 			WHERE MARKING_PERIOD_ID='" . UserMP() . "'